Single quote the database owner
There is an issue with our current version of PSQL (11.x) and this
library.

We're trying to create a database using this library and the statement
that this library use seems to be invalid with what PSQL expects.

From our experience, the CREATE DATABASE statement owner's parameter has
to either be quoted such that it is:

"CREATE DATABASE appraisal OWNER = 'XX_USERNAME_XX' ENCODING = 'UTF-8' TEMPLATE template1"

or that it's written without the "equal" sign such as

"CREATE DATABASE appraisal OWNER XX_USERNAME_XX ENCODING = 'UTF-8' TEMPLATE template1"

Otherwise, we get errors on the PSQL database:

```
2020-01-20 18:28:32 UTC:10.2.1.125(51484):XX_USERNAME_XX@postgres:[32587]:ERROR: role "XX_USERNAME_XX" does not exist
2020-01-20 18:28:32 UTC:10.2.1.125(51484):XX_USERNAME_XX@postgres:[32587]:STATEMENT: CREATE DATABASE appraisal OWNER = XX_USERNAME_XX ENCODING = 'UTF-8' TEMPLATE template1
```

The documentation of PSQL seems to describe that it should work with the equal sign, but our experiences tells us otherwise.

https://www.postgresql.org/docs/11/sql-createdatabase.html